Only cache objects managed by operator
Configure controller-runtime cache to only cache objects that have the commonly used labels set by the operator.
With that the operator will build up a cache for only the resources it manages.

Normally controller-runtime would build up a cache for all resources the client interacts with.
This is especially a problem for large cluster as then the operator would cache all Deployments and DaemonSets, which would then drastically increase memory usage.

With this change we can keep memory usage relatively small and also stable regardless the size of the cluster

Signed-off-by: Benjamin Isinger <Benjamin.Isinger@ibm.com>